requirements.txt
pytesseract
tabulate
pillow

